Extract methods and constants and add date params for fetching
