+++ openfemaapiclient/__init__.py
@@ -1,4 +1,4 @@
 from .core import fetch_from_api_generator, fetch_from_api, fetch_disaster_declarations, fetch_pa_applicants, fetch_pa_funded_projects
 from .disaster_declaration import DisasterDeclaration
-from .pa_applicant import PAApplicant
-from .pa_funded_project import PAFundedProject
+from .public_assistance_applicant import PAApplicant
+from .public_assistance_funded_project import PAFundedProject
