+++ openfemaapiclient/public_assistance_applicant.py
@@ -1,18 +1,17 @@
 from dataclasses import dataclass
 from datetime import datetime
+from .dateutils import try_extract_date
+
+DATETIME_FIELDS = {"lastRefresh"}
 
 
 def applicant_mapper(records):
-    applicant_datetime_fields = {"lastRefresh"}
+
     results = []
     for record in records:
-        mapped_result = {k: __to_datetime(v) if k in applicant_datetime_fields else v for k, v in record.items()}
+        mapped_result = {k: try_extract_date(v) if k in DATETIME_FIELDS else v for k, v in record.items()}
         results.append(PAApplicant(**mapped_result))
     return results
-
-
-def __to_datetime(value):
-    return datetime.strptime(value, "%Y-%m-%dT%H:%M:%S.%fZ") if value else None
 
 
 @dataclass(frozen=True)
